[level 0] Title: 양꼬치, Time: 0.04 ms, Memory: 73.7 MB -BaekjoonHub

+++ 프로그래머스/0/120830. 양꼬치/양꼬치.java
@@ -2,7 +2,7 @@ class Solution {
     public int solution(int n, int k) {
         int answer = 0;
         int drink = 0;
-        drink = k- (int)n/10;
+        drink = k- n/10;
         answer = n * 12000 +  drink * 2000;
         return answer;
     }
